boot: move purgatory to its own package
This prepares us for using a common purgatory between Linux and
multiboot.
