+++ pkg/boot/purgatory/asm.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-package linux
+package purgatory
 
 var Purgatories = map[string]*Purgatory{
 	"to32bit_3000": {
